:memo: :: 변수들 값 바뀌게 코드 변경

+++ app/src/main/java/com/g3c1/oasis_android/feature_select_store/data/datasourceimpl/StoreDataSourceImpl.kt
@@ -21,7 +21,7 @@ class StoreDataSourceImpl @Inject constructor(
 
         return flow {
             try {
-                val response = api.getStore(serialNumber = searialNumber)
+                val response = api.getStore(serialNumber = searialNumber.toLong())
                 if (response.isSuccessful) {
                     response.body()?.let {
                         emit(ApiState.Success(it))
